feat 28 - code review on data-processing-agreement
